Add static modifier to nested kotlin classes
Nested classes in kotlin are static by default.
This CL updates KspHasModifiers to add the static
modifier when it is a nested kotlin class without the
inner modifier.

A possible alternative solution is to add `isInner` to XTypeElement
because Room is using isStatic as a proxy to discover inner classes.
I've not implemented that because removing isStatic is not very feasible
from XHasModifiers but we can do a followup CL to divide XHasModifiers
and remove isStatic from there.

Right now, KSP does not report INNER modifier for .class files which is
why I couldn't add a test with compiled code.
https://github.com/google/ksp/pull/232

I've also renamed KspTypeElementTest to XTypeElementTest since
it runs all but 1 test w/ both processors.

Bug: 160322705
Test: XTypeElementTest
Change-Id: Ie07382a8744b26eda3b4454d45ea74bfae71fd70