Use AnnotatedString for EditCommands
This CL updates the EditCommand API to use AnnotatedString,
however it does not provide the multi style text editing
capabilities.

The goal is to define the AnnotatedString in the API and
implement the functionality later.

- EditCommand implementations accept AnnotatedString
- EditingBuffer accepts replace(AnnotatedString)
- Similar to aosp/1544024
- Kept the same API structure (added annotatedString)

Test: Manual test in the demo app
Test: ./gradlew compose:ui:ui-text:test
Test: ./gradlew compose:ui:ui-text:connectedAndroidTest
Test: ./gradlew compose:ui:ui:test
Test: ./gradlew compose:ui:ui:connectedAndroidTest

RelNote: "EditCommand’s accept AnnotatedString. However this is
an API only change and multi-style text editing is not
implemented yet."

Bug: 176815465
Change-Id: I4c3eaf53926b558759da95d498a07bc636ba4336