Early add Popup to window
When creating an AndroidPopup, don't wait until variables necessary to
calculate its position are available before adding the PopupLayout to
the WindowManager, but add it immediately.

To avoid flickering, the popup will be fully transparent while we don't
have all variables yet that are necessary to calculate the popup's
position.

Bug: 175765614
Test: Ran demos
Change-Id: I5b364106d86d79b2856ef72701f0307542676442