Fix BiometricManager error code for unsecured device
Currently, BiometricManager#canAuthenticate(int) will always return
BIOMETRIC_NONE_ENROLLED if the device is not secured with a credential,
since no form of authentication is possible. However, this error code
is incorrect if (1) the argument passed to canAuthenticate(int) does not
allow device credential authentication and (2) the device has no
biometric sensor. Instead, BIOMETRIC_ERROR_NO_HARDWARE should be
returned in this case.

This commit fixes the issue by returning BIOMETRIC_NONE_ENROLLED early
*only* if the device is not secured with a credential and device
credential authentication is allowed. In order to still indicate that
authentication is definitely not possible if the device is unsecured,
the canAuthenticate(int) method now returns the FingerprintManager error
code instead of BIOMETRIC_STATUS_UNKNOWN in some cases.

Test: Biometric integration test app on API 27-30.
Test: ./gradlew biometric:biometric:test
Test: ./gradlew biometric:biometric:connectedAndroidTest

Fixes: 174505824
Change-Id: I79b7dbf009c44582c4693be5d27dd850777ab4d1