Remove KeyInput modifier from FocusModifier
Currently every focus modifier has a keyinput modifier that captures keystrokes and moves focus based on the key pressed. Due to this
focusable components end up consuming the arrow key and tab keystrokes.
This CL removes the keyinput modifier from the focus modifier so that
users can receive all keyevents even if the component is focusable.

This logic is now moved to the owner, to the root keyinput modifier.

Bug: 177934607
Test: ./gradlew compose:ui:ui:connectedCheck -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.focus.CustomFocusTraversalTest
Change-Id: Iaeac1cdacf8b6b5183ebe855ac6dafd00396f6ef