Add methods to handle Ongoing Activity Data in Bundles.
Add two methods, one to copy information of an Ongoing Activity data
from one Bundle to another, and another to create an OngoingActivityData
instance from a Bundle.
This is useful so that WCS can pass on the Ongoing Activity Data with no
deserialization/serialization, which could result in data loss.

Test: Updated.
Relnote: N/A
Change-Id: I688d4b9a6559544c1635c80eb20b5dfb6d020790