Adds enabled parameters to Tab and BottomNavigationItem
Also changes BottomNavigationItem back to being an extension on RowScope.

BottomNavigationItems are equally spaced within a `BottomNavigation`, so by adding `RowScope` we can express this contract as part of the API. We previously removed this but having this contract makes it a lot harder to misuse the component and run into strange cases where the weight is not applied.

Also adds `Tab` role to BottomNavigationItem for accessibility purposes.

Fixes: b/152764416
Test: updateApi
Relnote: "Added enabled parameters to Tab and BottomNavigationItem, to prevent them from being clickable. Changed BottomNavigationItem to be RowScope.BottomNavigationItem to better express its layout requirements in its API."
Change-Id: Id683dc9a558c8d0704306a1c7f307985a9dbb12f