Ensure FragmentContainerView handles animators
When removing views that have animators, the view system puts exiting
views on top of entering views. FragmentContainerView was designed to
reverse the drawing order for at least animations and until recently
also handled animators.

The old animation system needed to check to see if the view had an
associated animation before considering it a disappearing view. The new
system does not require such a check.

RelNote: "Animators now appear in the proper order when using
FragmentContainerView"
Test: FragmentContainerViewTest
Bug: 176089197

Change-Id: Id9aa349b180cc2c22912efae3f2c44773ecb7126