Make all model getters return nullable except if they are collections.
This shifts the responsibility of null-checking to the host code that
consumes the library. This approach is more defensive and future-proof
if and when we need to change any required fields in the models in future
versions of the library.

This practice is also helpful for pre-beta releases of the library where
we previously have switched the nullability on the getters, resulting in
incompatible host(s). We have run into issues in the past where we need
to do another release because of a null-check change.

Relnote: n/a
Bug: 178103614
Bug: 175308069
Test: ./gradlew car:app:app:test
Change-Id: I31b7cdb3947f22182cf401c0e0f882e9b13e4ed8