Introduce SnapshotMutableState interface
It is a simple interface which extends MutableState for the implementations where writes are transacted as part of the Snapshot system. This interface also defines a SnapshotMutationPolicy property.
This change will allow us to automatically save/restore objects created via mutableStateOf() via saved instance state mechanism

Relnote: SnapshotMutableState was introduced
Test: ./gradlew bOs
Change-Id: Icfd03593afd98c2cd1ad94751e590f4eefbc06e5