Set shutdown ListenableFuture cannot be cancelled
The returned ListenableFuture of the CameraX.shutdown() might
be cancelled unexpectedly when the shutdown task is timeout,
but the internal clean up tasks for shutdown don't stop. It would
cause the initialization task to run with the clean up task
in parallel, and result in the incorrect internal state.
Wrap the returned ListenableFuture with the nonCancellationPropagating
ListenableFuture to keep the internal shutdown ListenableFuture
cannot be cancelled.

Bug: 178127576
Test: manually test the CameraXTest
Change-Id: I2428df288248b96c56a9d96f7665cc6248d2698f