Introduce WatchFace.getOrCreateEditorDelegate
Previously we assumed an on watch face editor would only be invoked
when the watch face is already loaded. This is incorrect on Wear OS
2.0 (see bug for details).  We now cope with the watch face being
constructed after the editor.

Test: Manual testing confirmed the bug has gone.
Bug: 178460984
Relnote: Please note createOnWatchEditingSession is now a suspended function because the watchface sometimes isn't avaiable until shortly after the editor activity has started.
Change-Id: Ida9aa6ee3801b9ff79f64966340fdf485f1af389