Keep track of active data store files so we can throw an exception if there is a duplicate instance.
Relnote: DataStore will now throw an exception if there are multiple active DataStores for the same file. If you were not managing your DataStore as a Singleton or were not ensuring that no two instances of DataStore are simultaneously active for a file then you may now see exceptions when reading or writing to DataStore. These can be fixed by managing your DataStore as a Singleton.
Bug: 177691248
Test: added tests to SingleProcessDataStoreTest
Change-Id: Ib43f4495f7741a3d480e0f6256e7bfee512c0d56