Add a minimal example watch face in Java.
This commit adds a minimal watch face implementation in Java. This is
useful both to provide a sample of how to use our APIs from Java
clientand validating the wrappers around suspend functions.

Being a really minimal implementation, this is also a good starting
point to test other features being developed.

This is set up as a separate project because we want to be able to
test the way we ask our developers to develop watch faces, which is
where each watch face is an individual APK. Therefore, we plan to
create a suite of these samples, showcasing the different options for
developers.

Bug: 179233274
Test: ./gradlew :wear:wear-watchface-samples-minimal:assembleDebug
Change-Id: I2626d3c160c15147edf9e26e10ad70cc02ea28bd