Run DaoWriter tests w/ KSP
I had to change the WriterDao expected output to not include copyright
because XTest code equality is not clever enough to do structural
comparison.

Also, some methods in test_utils that work with TestInvocation are
duplicated to also work with XTestInvocation. Eventually, once all tests
move, we'll be able to remove all JavaFileObject access in the codebase.

Bug: 160322705
Test: DaoWriterTest
Change-Id: Ica1bbf265768b9bd93165602b89db54496b186fb