[GH] Add initial infrastructure for FragmentStrictMode
## Proposed Changes

  - Add initial infrastructure for `FragmentStrictMode`

Found this in the issue tracker and think it's a great idea. I tried to keep the first version as lean and as closely aligned to `StrictMode` as possible. I also plan on creating some follow-up PRs to add a few checks (see issue tracker), as soon as the infrastructure is discussed and somewhat agreed upon. For now the classes are restricted to the library, until we have the first few checks.

## Testing

Test: See `FragmentStrictModeTest`

## Issues Fixed

Fixes: 153737341

This is an imported pull request from https://github.com/androidx/androidx/pull/123.

Resolves #123
Github-Pr-Head-Sha: e79ad8e4c100f12e637877e5cc6c0764a11de680
GitOrigin-RevId: 5c1aa683d57c18c300f5e4c21dcffd219ac8d675
Change-Id: Ib634f058efddf686db3de6be0dc6b7af1a3de718