Fixes initial PopupDemo intercepting clicks / back press and not dismissing
The initial popup demo was focusable and (by design) didn't dismiss itself on click outside / back press, so it could not be navigated away from. Now this popup is not focusable, so it will not intercept back presses and touches will still reach content underneath after the popup sees them.

Fixes: b/179339732
Fixes: b/180493147
Test: DemoTest
Change-Id: I15ab07ad9dbc0d327d4f56f270a6f44c47e2793b