Fix inflating <fragment> tag without a container
aosp/1591056 fixed an issue that was preventing <fragment> tag inflated
fragments from getting to resume if the activity was already resumed. It
used the container passed in from inflate, but that can be null as in
the case of inflating a view for use with the dialog fragment, so we
should use fragment view's parent. We know this will not be null since
it is inside of an onViewAttachedtoWindow listener.

Test: testInflatedDialogFragmentShows
Bug: 180538371
Fixes: 180637744
Change-Id: I11920da0b6603db82b272edd553f8e66b3c9248c