Ensure inflated fragments that are recreated are shown
When using the new fragment state manager, if a fragment is inflated
after the host activity is RESUMED using a FragmentContainerView, the
fragment is added and correctly moves to RESUMED the first time, but
upon config change, the fragment's view is never readded to the
container. This results in the fragment moving to RESUMED but never
being reattached to the view hierarchy.

This change finds the restored fragment and indicates to the
FragmentManager that the FragmentContainerView is ready for the fragment
to be attached.

RelNote: "Fragments inflated into an already RESUMED activity using
FragmentContainerView are now properly shown after a configuration
change."
Test: DialogFragmentTest and FragmentLifecycleTest
Bug: 180538371

Change-Id: Ie14c80467b66186862319469f51bbd4d7ebcbbb9