Avoid redundantly saving direct boot params
This is an optimzation, we were saving the params every time the
UserStyle was assigned including when taking screen shots (which
set a new style and restored the orgional) regardless of whether
the style actually changed or not. Writing to storage is not free
so we don't want to update this unless actually required.

Bug: 180913774
Test: Manual testing with perfetto
Change-Id: I5f246ab59132fdc42ed28282625e89e63fe3fdab