Re-trigger access hint when inserted page is empty
This CL fixes a bug in paging where if an inserted page is empty, we
wouldn't send another access hint to load more and RecyclerView would
also not send any hints since it won't be updated.

With this change, if the inserted page is empty, we trigger another
access hint to load more.

Bug: 180691012
Test: LoadFullListTest, sample app in the bug
Change-Id: I15feb352d088799c255cd1963dbca38dec46e7fc