Mark LayoutInspectorTreeTest with @MediumTest
Mark this test as @MediumTest to make sure it runs in presubmit. Even though
testStitchTreeFromModelDrawerLayoutWithSystemNodes runs in less than 1s (the
@MediumTest annotation threshold) both locally and remotely on Sponge 2,
it's throwing TestTimedOutException remotely, probably because the test
timings are not too reliable and it's close to the threshold (> 900ms). For
that reason, we explicitly mark this individual test as @LargeTest.

Bug: 179770455
Test: LayoutInspectorTreeTest
Change-Id: I952a870c47fa69a39daef06d9adc2556b8fcba56