Handling new table addition schema changes.
In the current implementation, the AutoMigration code will be able to detect the addition of a new table to a schema. It will also be able to detect if a table has been deleted (or renamed, can't difference between these two in this implementation). In this case, the processor will output an error message that includes the name of the table in reference. In the follow-up CLs to this CL, we will begin handling complex schema changes.

Test: Added testing for new column addition in valid and invalid cases for the AutoMigrationProcessorTest.kt, AutoMigrationWriterTest.kt, SchemaDifferTest.kt
Bug: 180119333
Change-Id: Ifcfaf8c118a340bb3a0a5e81a0fedc8b797c2273