Optimize watch face coroutine dispatch
Previously continuations on the same thread where posted on the
handler. This is problematic because the main thread can be busy,
delaying work e.g. during init (sometimes it made it take 2x as
long). In addition we add some new trace events for watchface init
and we fully fix construction of a headless instance by ensuring
maybeCreateWCSApi won't run.

Bug: 181366521
Test: Manual testing
Change-Id: Ia0fb7c97d5bbeeb85fd0a7b390aff50edcee4fc8