Make LocalViewModelStoreOwner.current Nullable
While in most cases a LocalViewModelStoreOwner is
automatically made available via the
ViewTreeViewModelStoreOwner API, it may not always
be available. By moving the error up from the
LocalViewModelStoreOwner layer to the APIs that
require it (namely, viewModel() and NavHost), other
components can use the nullability of the
LocalViewModelStoreOwner as an indication on whether
these APIs can be used.

Relnote: "`LocalViewModelStoreOwner.current` now returns a
nullable `ViewModelStoreOwner` to better determine
whether a `ViewModelStoreOwner` is available in the current
composition. APIs that require a `ViewModelStoreOwner`, such
as `viewModel()` and `NavHost`, still throw an exception
if a `ViewModelStoreOwner` is not set."
Test: new nullViewModelStoreOwner test

Change-Id: Idf39af28f00e0c0522a897a1202385d21da828be