Transition-based AnimatedVisibility
This CL improved AnimatedVisibility by putting all the enter/exit
animations in a Transition. As a result these animations are no longer
completely internal to AnimatedVisiblity, but instead their lifecycle
can be observed from outside.

To achieve this, there are a few fundamental changes done to Transition:
1. Added support for deferred animations, so that Transition animations
   can be initialized after composition stage (e.g. during measure pass).
   This made it possible to re-create the slide/shrink/expand animations
   that are size
   dependent.
2. New `createChildTransition` API to allow child Transitions added to
   a parent Transition of different type. This enabled the
   Transition<EnterExitState> that AnimatedVisibility creates internally
   to be hoisted into another Transition object, and therefore have their
   lifecycle observed.

These changes in Transition now allows users to create additional
animations as a part of enter/exit transition, such that
AnimatedVisibility will factor in those custom animations in determining
its state. This was previously not possible.

Another improvement made to AnimatedVisibility is the support of
MutableTransitionState. This not only solves the confusion around the
`initiallyVisible` parameter, but also simplifies the data sync between
UI and Model. More specifically, business logic can influence
`targetState` and therefore triggering animation, which has its state
reflected in `currentState`. Once the animation is finished, business
logic can observe it and follow up with data updates as needed.

Also introduced AnimatedVisibilityScope to support:
1. Use of default enter/exit animations in children via
   Modifier.animateEnterExit
2. Custom enter/exit animation via #transition

Fixes: 163810093
Test: Added
RelNote: "
- New AnimatedVisibility API that support visibility to be specified in
a MutableTransitionState. This also allows the animation states to be
obseved external to AnimatedVisibility via currentState and isIdle.
- New child transition support in Transition
- Support custom exit/enter transition animation in AnimatedVisibility
  using AnimationScope.transition
- New animateEnterExit modifier accessible for all children of
  AnimatedVisibility
- Deprecated the AnimatedVisibility APIs that take an `initiallyVisible`
  parameter
"

Change-Id: I702f3a6d948938130701e6ab91b77f0b6fe000c2