Lint check to warn when directly creating coroutines inside a composable
Instead developers should use a `LaunchedEffect` to ensure that new coroutines are not created if composition fails or is rolled-back, and to ensure that they are cancelled correctly when inputs change.

Also stops ReportLibraryMetricsTask from checking `lint.jar` - these aren't part of the library so no need to check them.

Test: ComposableCoroutineCreationDetectorTest
Fixes: b/180729454
Change-Id: I53200a24284c94e695562625836d14c1fb0f7c0b