Changes some correctness compose:ui lint checks to be warnings, not errors
Also softens the 'must' wording to be 'should' for ModifierDeclarationDetector

We still want these to be errors in our build, as these are blocking issues for internal code, so set them to be an error in AndroidXUiPlugin.

Additionally, disables UnknownIssueId and removes existing usages from baselines. Unfortunately there is a bug (182321297) where ignoring or disabling this issue still prints some messages to the console, so this Cl also updates messages.ignore to account for the new message.

Bug: b/173708494
Test: existing tests
Change-Id: Ia8a88ccf26778b4d7b17092e5d593d0fd0cd65d8