Support multiple listeners in SlidingPaneLayout
Rather than only allowing a single PanelSlideListener
via setPanelSlideListener(), replace that API with
a set of add/remove methods that allow for developers
to implement separate, independent listeners.

Use this new functionality in the support-v4-demos
app to demo a reusable SliderOnBackPressedCallback
that intercepts the system back button and closes
the SlidingPaneLayout if it is open.

Test: new PanelSlideListenerTest
Relnote: "SlidingPaneLayout now allows registering
multiple PanelSlideListeners."

Change-Id: I50ce2f92348cc3c9e6ee8a7e27ae64e62e88e88f