Fix testAmbientScreenshot flakes
Upon instrumenting it became apparant that the code setting the
ComplicationRenderer's ambient was being applied to the wrong
ComplicationRenderer (which gets reassigned after
ComplicationDrawable's init). This change should be a nop but it's
not suggesting either a misunderstanding of Kotlin or a compiler
bug.

Test: Manual testing
Relnote: Changed the name of a CanvasComplicationDrawable parameter to fix a bug.
Change-Id: I50dac9dd98c1461e44427134813d1866f1476f94