Update the Compose Compiler plugin version checker for beta04
This should have landed with the bump to 1.0.0-beta04

This also aligns androidx.hilt:hilt-navigation-compose
to use the same version of compose and kotlin. Furthermore
it also updates navigation-compose to not pin its dependency
to viewmodel-compose since that will downgrade the runtime
on Android Test variants due to AGP constrains.

Bug: 183511654
Test: gw bOS
Test: gw :hilt:hilt-navigation-compose:compileReleaseKotlin
Test: gw :navigation:navigation-compose:compileDebugAndroidTestKotlin
Change-Id: I8e79af3f1a27d0c2adaa3ad80e1666be8ded2027