[cxcp-3a-bridge-2] Extend Result 3A
Add 3A state and metering regions to Result3A. This helps the clients
to know about the state of 3A state machine after a 3A operation is
complete.
In camerax for example, we check if the focus state was exactly
focused_locked and metering regions were approximately the same,
in order to set the status as 'true' for MeteringResult. In CameraPipe
we care about whether the command was submitted and the result has one
of the valid states for 3A.

Extending the Result3A will help us to add more filters at the adapter
layer.

Test: ./gradlew :camera:camera-camera2-pipe:test
      ./gradlew :camera:camera-camera2-pipe-integration:test

Change-Id: I8954a7fbad4293d54735502bd1acb1235a1daa58